requirements.txt
streamlit
langchain
google-generativeai
python-dotenv
nltk
arxiv
langchain_google_genai